feat(display): Hardware initialization function
Create the eInkDisplay_Init which is a funtion that will initilaze the
display Hardware. This function is not finished yet, it just contains
the peripheral initializations using two static functions, one for GPIO
initialization and the other for SPI.

JIRA: SCRUM-20
