+++ user/Inc/main.h
@@ -1,4 +1,5 @@
 #include "stm32f429zi.h"
+#include "einkPaper_2_13.h"
 
 
 void Test(void);
